Experimentally add capnp implementation
In order to compare it directly with the grpc implementation, here is the implementation using capnproto.

TBH, the developer experience is equally as bad, or even worse.
Similarly, one needs to have `protoc` installed (for good reasons).
The generated code hardcodes some `crate` paths for reasons I don’t understand.
Plus, capnp brings its own `Promise` abstraction which is just a worse `Future`, and requires pretty much everything to be wrapped in `pry!` macros.
It is also weird that the response is written into an out-parameter. Its not obvious what happens to it when an error is being returned.